moved scicrunch utils
